fix naming convention and split tests (#1536)
* fix naming convention and split tests

* remove moved tests
