+++ projects/storefrontapp-e2e-cypress/cypress/integration/regression/added-to-cart-modal.e2e-spec.ts
