+++ projects/storefrontapp-e2e-cypress/cypress/integration/regression/address-book.e2e-spec.ts
@@ -1,7 +1,7 @@
 import {
   fillShippingAddress,
   ShippingAddressData
-} from '../helpers/checkout-forms';
+} from '../../helpers/checkout-forms';
 
 describe('Address management page', () => {
   const newAddress: ShippingAddressData = {
