+++ projects/storefrontapp-e2e-cypress/cypress/integration/regression/product-details.e2e-spec.ts
