+++ projects/storefrontapp-e2e-cypress/cypress/integration/regression/product-search/product-search-rating-flow.e2e-spec.ts
@@ -1,4 +1,4 @@
-import { PRODUCT_LISTING } from './../helpers/data-configuration';
+import { PRODUCT_LISTING } from '../../../helpers/data-configuration';
 context('Product search rating flow', () => {
   before(() => {
     cy.visit('/');
