+++ projects/storefrontapp-e2e-cypress/cypress/integration/regression/store-finder.e2e-spec.ts
