+++ projects/storefrontapp-e2e-cypress/cypress/integration/smoke/homepage.e2e-spec.ts
