+++ projects/storefrontapp-e2e-cypress/cypress/integration/smoke/login.e2e-spec.ts
@@ -1,5 +1,5 @@
-import { user } from '../sample-data/big-happy-path';
-import { register, login } from '../helpers/auth-forms';
+import { user } from '../../sample-data/big-happy-path';
+import { register, login } from '../../helpers/auth-forms';
 
 describe('Login', () => {
   const userGreetSelector = 'cx-login .cx-login-status__greet';
