+++ projects/storefrontapp-e2e-cypress/cypress/integration/smoke/routing.e2e-spec.ts
